AsyncFunctionToPassportFirebase
